+++ resources/lang/en/addon.php
@@ -3,5 +3,5 @@
 return [
     'title'       => 'PyroCMS Plus',
     'name'        => 'PyroCMS Plus Theme',
-    'description' => 'PyroCMS Plus Theme'
+    'description' => 'PyroCMS Plus Admin Theme'
 ];